// types/leadInfoTypes.ts
export type LeadInfoPayloadTypes = {
  id: number;
  name: string;
  dob: string | null;
  age: string | null;
  mobile_num: string | null;
  email_address: string | null;
  spouse_dob: string | null;
  spouse_age: string | null;
  kids: string | null;
  family_kid1_dob: string;
  family_kid2_dob: string | null;
  family_kid3_dob: string | null;
  family_kid4_dob: string | null;
  product_type: string | null;
  product_data: string | null;
  product_sku: string | null;
  travel_from: string | null;
  travel_to: string | null;
  travel_go: string | null;
  travel_with: string | null;
  travel_country: string | null;
  from_date: string | null;
  end_date: string | null;
  duration: string | null;
  insure_against: string | null;
  insure_house: string | null;
  insure_for: string | null;
  insure_live: string | null;
  insure_my: string | null;
  insure_area: string | null;
  insure_live_in: string | null;
  parent_insurance: string | null;
  parents_or_inlaw: string | null;
  parents_mother_dob: string | null;
  parents_father_dob: string | null;
  status: string | null;
  created_by: number;
  is_active: boolean;
  is_deleted: boolean;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
};

export type LeadInfoResponseTypes = {
  status: 1 | 0;
  message: string;
  payload: LeadInfoPayloadTypes[];
};